voice_client/requirements.txt
# Voice Assistant Requirements

# Core dependencies
websockets>=12.0
python-dotenv>=1.0.0
xai-sdk>=0.1.0

# Audio I/O (required for microphone and speaker)
pyaudio>=0.2.13

# Optional: for better async performance
aiohttp>=3.9.0
